refactor(github/vulnerability): use Zod-inferred types as single source of truth
