+++ lib/modules/platform/github/types.ts
@@ -7,6 +7,14 @@ import type { Pr, PrBodyStruct } from '../types.ts';
 export type CombinedBranchState = 'failure' | 'pending' | 'success';
 export type BranchState = 'failure' | 'pending' | 'success' | 'error';
 
+type VulnerabilityKey = string;
+type VulnerabilityRangeKey = string;
+type VulnerabilityPatch = string;
+export type AggregatedVulnerabilities = Record<
+  VulnerabilityKey,
+  Record<VulnerabilityRangeKey, VulnerabilityPatch | null>
+>;
+
 export interface GhBranchStatus {
   context: string;
   state: BranchState;
