+++ lib/modules/platform/types.ts
@@ -1,19 +1,9 @@
 import type { DateTime } from 'luxon';
 import type { MergeStrategy } from '../../config/types.ts';
-import type {
-  BranchStatus,
-  HostRule,
-  VulnerabilityAlert,
-} from '../../types/index.ts';
+import type { BranchStatus, HostRule } from '../../types/index.ts';
 import type { CommitFilesConfig, LongCommitSha } from '../../util/git/types.ts';
-
-type VulnerabilityKey = string;
-type VulnerabilityRangeKey = string;
-type VulnerabilityPatch = string;
-export type AggregatedVulnerabilities = Record<
-  VulnerabilityKey,
-  Record<VulnerabilityRangeKey, VulnerabilityPatch | null>
->;
+import type { GithubVulnerabilityAlert } from './github/schema.ts';
+export type VulnerabilityAlert = GithubVulnerabilityAlert;
 
 export interface PlatformParams {
   endpoint?: string;
