--- lib/modules/platform/utils/github-alerts.ts
@@ -1,14 +1,0 @@
-import type { VulnerabilityPackage } from '../../../types/index.ts';
-import { normalizePythonDepName } from '../../datasource/pypi/common.ts';
-
-export function normalizeNamePerEcosystem({
-  name,
-  ecosystem,
-}: VulnerabilityPackage): string {
-  switch (ecosystem) {
-    case 'pip':
-      return normalizePythonDepName(name);
-    default:
-      return name;
-  }
-}
