+++ lib/types/index.ts
@@ -22,12 +22,6 @@ export type { CombinedHostRule, HostRule } from './host-rules.ts';
 export type { PrState } from './pr-state.ts';
 export type { SkipReason, StageName } from './skip-reason.ts';
 export type { RangeStrategy } from './versioning.ts';
-export type {
-  SecurityAdvisory,
-  SecurityVulnerability,
-  VulnerabilityAlert,
-  VulnerabilityPackage,
-} from './vulnerability-alert.ts';
 
 export type AutoMergeType = 'branch' | 'pr' | 'pr-comment';
