+++ lib/workers/repository/init/vulnerability.ts
@@ -5,6 +5,7 @@ import { GithubTagsDatasource } from '../../../modules/datasource/github-tags/in
 import { GoDatasource } from '../../../modules/datasource/go/index.ts';
 import { MavenDatasource } from '../../../modules/datasource/maven/index.ts';
 import { NugetDatasource } from '../../../modules/datasource/nuget/index.ts';
+import type { SecurityAdvisory } from '../../../modules/platform/github/schema.ts';
 import { platform } from '../../../modules/platform/index.ts';
 import * as composerVersioning from '../../../modules/versioning/composer/index.ts';
 import * as allVersioning from '../../../modules/versioning/index.ts';
@@ -13,7 +14,6 @@ import * as npmVersioning from '../../../modules/versioning/npm/index.ts';
 import * as pep440Versioning from '../../../modules/versioning/pep440/index.ts';
 import * as rubyVersioning from '../../../modules/versioning/ruby/index.ts';
 import * as semverVersioning from '../../../modules/versioning/semver/index.ts';
-import type { SecurityAdvisory } from '../../../types/index.ts';
 import { sanitizeMarkdown } from '../../../util/markdown.ts';
 import { escapeRegExp } from '../../../util/regex.ts';
 import { githubEcosystemToDatasource } from '../../../util/vulnerability/ecosystem.ts';
